FIPS: Move Algorithm class into a separate file.
